Add DealDamage state, guard physics until weapon and enemy are chosen
Also switch to uint8_t/uint16_t to saves memory with <cstdint> and comment out the old terminal loop

+++ CombatUtils.h
@@ -1,7 +1,8 @@
 #include <iostream>
+#include <cstdint>
 #pragma once
 
-inline void ApplyDamage(int& health, int& damage,bool& isAlive){
+inline void ApplyDamage(int& health, u_int8_t& damage,bool& isAlive){
     health -=damage;
     if (health<=0){
         isAlive=false;
