+++ Player.h
@@ -3,6 +3,7 @@
 #include "weapon.h"
 #include <vector>
 #include <utility>
+#include <cstdint>
 
 class Enemy;
 
@@ -10,7 +11,7 @@ class Player {
     public:
         std::string name = "Link";
         int health = 100;
-        int maxHealth = 100;
+        uint8_t maxHealth = 100;
         bool isAlive = true;
         bool isReturn = false;
         bool isAttack = false;
@@ -18,8 +19,8 @@ class Player {
         bool isKnocked = false;
         float posX = 150;
         float posY = 300;
-        int basePosX = 150;
-        int basePosY = 300;
+        u_int16_t basePosX = 150;
+        u_int16_t basePosY = 300;
         float vX = 0;
         float vY = 0;
         float NormalizedHypotenous = 0;
@@ -27,7 +28,7 @@ class Player {
         Enemy* TargetEnemy = nullptr;
         std::vector<std::pair<float, float>>Recall;
     public:
-        void TakeDamage(int damage);
+        void TakeDamage(u_int8_t damage);
         void Attack();
         void Return();
         void KnockBack(float VxOpponent, float VyOpponent);
