+++ weapon.cpp
@@ -2,7 +2,7 @@
 #include <iostream> 
 
 //ClassName::FunctionName is the syntax to define fx more .Constructor
-Weapon::Weapon(std::string weaponName,int weaponDamage, int weaponDurability) : name(weaponName), damage(weaponDamage), durability(weaponDurability) {
+Weapon::Weapon(std::string weaponName,u_int8_t weaponDamage, u_int8_t weaponDurability) : name(weaponName), damage(weaponDamage), durability(weaponDurability) {
 }
 
 void Weapon::Attack() {
